refactor(config): move config files to their modules
