+++ src/app.module.ts
@@ -2,15 +2,15 @@ import { Module } from '@nestjs/common';
 import { UsersModule } from './users/users.module';
 import { FilesModule } from './files/files.module';
 import { AuthModule } from './auth/auth.module';
-import databaseConfig from './config/database.config';
-import authConfig from './config/auth.config';
+import databaseConfig from './database/config/database.config';
+import authConfig from './auth/config/auth.config';
 import appConfig from './config/app.config';
-import mailConfig from './config/mail.config';
-import fileConfig from './config/file.config';
-import facebookConfig from './config/facebook.config';
-import googleConfig from './config/google.config';
-import twitterConfig from './config/twitter.config';
-import appleConfig from './config/apple.config';
+import mailConfig from './mail/config/mail.config';
+import fileConfig from './files/config/file.config';
+import facebookConfig from './auth-facebook/config/facebook.config';
+import googleConfig from './auth-google/config/google.config';
+import twitterConfig from './auth-twitter/config/twitter.config';
+import appleConfig from './auth-apple/config/apple.config';
 import path from 'path';
 import { ConfigModule, ConfigService } from '@nestjs/config';
 import { TypeOrmModule } from '@nestjs/typeorm';
