+++ src/auth-apple/config/apple-config.type.ts
@@ -1,0 +1,3 @@
+export type AppleConfig = {
+  appAudience: string[];
+};
