+++ src/auth-facebook/config/facebook-config.type.ts
@@ -1,0 +1,4 @@
+export type FacebookConfig = {
+  appId?: string;
+  appSecret?: string;
+};
