+++ src/auth-google/config/google-config.type.ts
@@ -1,0 +1,4 @@
+export type GoogleConfig = {
+  clientId?: string;
+  clientSecret?: string;
+};
