+++ src/auth-google/config/google.config.ts
@@ -1,23 +1,23 @@
 import { registerAs } from '@nestjs/config';
-import { FacebookConfig } from './config.type';
+import { GoogleConfig } from 'src/auth-google/config/google-config.type';
 import { IsOptional, IsString } from 'class-validator';
 import validateConfig from 'src/utils/validate-config';
 
 class EnvironmentVariablesValidator {
   @IsString()
   @IsOptional()
-  FACEBOOK_APP_ID: string;
+  GOOGLE_CLIENT_ID: string;
 
   @IsString()
   @IsOptional()
-  FACEBOOK_APP_SECRET: string;
+  GOOGLE_CLIENT_SECRET: string;
 }
 
-export default registerAs<FacebookConfig>('facebook', () => {
+export default registerAs<GoogleConfig>('google', () => {
   validateConfig(process.env, EnvironmentVariablesValidator);
 
   return {
-    appId: process.env.FACEBOOK_APP_ID,
-    appSecret: process.env.FACEBOOK_APP_SECRET,
+    clientId: process.env.GOOGLE_CLIENT_ID,
+    clientSecret: process.env.GOOGLE_CLIENT_SECRET,
   };
 });
