+++ src/auth-twitter/config/twitter-config.type.ts
@@ -1,0 +1,4 @@
+export type TwitterConfig = {
+  consumerKey?: string;
+  consumerSecret?: string;
+};
