+++ src/auth/config/auth-config.type.ts
@@ -1,0 +1,10 @@
+export type AuthConfig = {
+  secret?: string;
+  expires?: string;
+  refreshSecret?: string;
+  refreshExpires?: string;
+  forgotSecret?: string;
+  forgotExpires?: string;
+  confirmEmailSecret?: string;
+  confirmEmailExpires?: string;
+};
