+++ src/auth/config/auth.config.ts
@@ -1,5 +1,5 @@
 import { registerAs } from '@nestjs/config';
-import { AuthConfig } from './config.type';
+import { AuthConfig } from 'src/auth/config/auth-config.type';
 import { IsString } from 'class-validator';
 import validateConfig from 'src/utils/validate-config';
