+++ src/config/app-config.type.ts
@@ -1,0 +1,11 @@
+export type AppConfig = {
+  nodeEnv: string;
+  name: string;
+  workingDirectory: string;
+  frontendDomain?: string;
+  backendDomain: string;
+  port: number;
+  apiPrefix: string;
+  fallbackLanguage: string;
+  headerLanguage: string;
+};
