+++ src/files/config/file-config.type.ts
@@ -1,0 +1,9 @@
+export type FileConfig = {
+  driver: string;
+  accessKeyId?: string;
+  secretAccessKey?: string;
+  awsDefaultS3Bucket?: string;
+  awsDefaultS3Url?: string;
+  awsS3Region?: string;
+  maxFileSize: number;
+};
