+++ src/mail/config/mail-config.type.ts
@@ -1,0 +1,11 @@
+export type MailConfig = {
+  port: number;
+  host?: string;
+  user?: string;
+  password?: string;
+  defaultEmail?: string;
+  defaultName?: string;
+  ignoreTLS: boolean;
+  secure: boolean;
+  requireTLS: boolean;
+};
